Fix CRUD operations for student exams and test runs, make sure they dont interfere with eachother

+++ src/main/java/de/tum/in/www1/artemis/web/rest/ExamResource.java
@@ -609,10 +609,14 @@ public class ExamResource {
 
             // Optionally delete participations and submissions
             if (withParticipationsAndSubmission) {
-                List<StudentParticipation> participations = participationService.findByStudentIdAndIndividualExercisesWithEagerSubmissionsResult(student.getId(),
-                        studentExam.getExercises());
-                for (var participation : participations) {
-                    participationService.delete(participation.getId(), true, true);
+                // check that no test runs reference these participations and submissions
+                List<StudentExam> testRuns = studentExamService.findAllTestRunsForUser(examId, student.getId());
+                if (testRuns.size() == 0) {
+                    List<StudentParticipation> participations = participationService.findByStudentIdAndIndividualExercisesWithEagerSubmissionsResult(student.getId(),
+                            studentExam.getExercises());
+                    for (var participation : participations) {
+                        participationService.delete(participation.getId(), true, true);
+                    }
                 }
             }
